fix(ner): align is_valid_ner_entity signature with Entity.count (usize)
The E0308 type mismatch at parse_md_full.rs:196 was caused by
is_valid_ner_entity being declared with count: u32 while Entity.count
in ner.rs is usize. Rust does not perform implicit numeric coercions,
so the call site failed to compile.

Fix: change the function signature to count: usize. This is the safer
choice compared to try_into().unwrap() because:
  - zero panic risk (no narrowing conversion)
  - matches the source-of-truth type (Entity.count: usize)
  - keeps the call site clean
  - the function only compares count against small thresholds (2/3/5),
    so u32 was never semantically required

Also clean up all 10 compiler warnings:
  - parse_md_full.rs: drop unused Entity import
  - models/mod.rs: drop unused re-exports (LitNodeType, EdgeKind, GraphData)
  - ai/mod.rs: drop unused AiResponse re-export
  - lib.rs: drop unused 'use tauri::Manager'
  - epsilon.rs: move kw_count inside the if-let block (kills unused_assignment
    + removes a spurious 'mut')
  - parser/mod.rs: drop .enumerate() in three loops where idx was unused

+++ src-tauri/src/commands/parse_md_full.rs
@@ -12,7 +12,7 @@
 //! добавляет недостающие имена, исправляет леммы, но не трогает chapters
 //! и locations (Rust-парсер для них достаточно точен).
 
-use crate::commands::ner::{run_python_with_text_file, NerResult, Entity};
+use crate::commands::ner::{run_python_with_text_file, NerResult};
 use crate::models::{ParseParams, ParseResult};
 use crate::parser;
 use serde::{Deserialize, Serialize};
@@ -121,7 +121,7 @@ fn run_ner_safe(text: &str) -> Option<NerResult> {
 ///   3. lemma содержит одновременно кириллицу И латиницу ("Root-Оператор")
 ///   4. lemma содержит цифры или спецсимволы (кроме дефиса и апострофа)
 ///   5. count < MIN_NER_COUNT (3 для PER, 5 для ORG)
-fn is_valid_ner_entity(lemma: &str, label: &str, count: u32) -> bool {
+fn is_valid_ner_entity(lemma: &str, label: &str, count: usize) -> bool {
     if lemma.is_empty() {
         return false;
     }
